Fix a potential deadlock when starting to profile
MVM_profile_instrumented_start waited on tc->instance->cond_spesh_sync without
marking the thread blocked. If the spesh thread was at that point waiting for
the other thread to join the GC run, we would end up in a dead lock.

Fix by marking the thread blocked while waiting for the mutex and cond var.