Fix search for bind control records
It is possible that the bind control record is in a stack segment prior
to the one that a frame is allocated in. In code that recurses enough to
trigger multiple stack segments, this situation may occur, and could
lead to multiple dispatch of candidates with `where` clauses and similar
wrongly reporting an error instead of continuing to the next candidate.