Always emit guards on resume init state
Even if there are literal flags in the capture, they were literal values
at the callsite of the *initial* dispatch, but can most certainly not be
assumed to be literal at the point we do a `callsame`. This led to the
generation of dispatch programs without literal string guards, which
would in turn lead to callsame in two different methods on the same type
but with different names going to the wrong place.