Prepare init resume args for translated DPs
When we translate a dispatch program into spesh ops, we'll also be
translating each resume init into an sp_resumption instruction, which
keeps the arguments and temporaries needed for resuption alive. This
means the way we access the resume init args is quite different from the
case where we had a dispatch recorded or dispatch run record on the call
stack.

This commit arranges for us to handle both (although for now the second
never occurs). We also insert holes into the list of registers indices
where there are constants in the resume initialization arguments, to
permit O(1) lookup of them.