Don't inappropriately copy facts when inlining
When we transform the `return` instruction, it makes sense to copy facts
when were are returning an object to a place where an object is wanted.
However, in the case that we are performing an unbox, they should not be
copied, otherwise we end up with object-y facts on a non-object operand,
which can in turn cause confusion (in the case debugged, it cause a
mis-optimization of a branch).