Better choice if proxy deopt index on inline
This is used to make sure we don't do changes that would impact our
ability to deoptimize inside of the inline and reconstruct the caller.
Prior to new-disp, we used the deopt index on the prepargs, and did
recreate that - well, sort of; in fact we produced a new index, and
achieved nothing at all. This fixes it by using a real deopt index with
deopt usages attached to it - but picks the deopt all index instead. The
reasoning is as follows: if we made it into an inline, that means all of
the guards stacked up ahead of it were successful, and so we'd never
deopt to the point before the overall dispatch. Thus it's the things
beyond the dispatch - including its return register - that matter, and
that we want to ensure have their deopt usages retained.