Correct deopt handling in runbytecode opts
When we stack up guards in front of a runbytecode so we can spesh link
or inline it, they need to have a deopt index allocated, and it needs to
be a predeopt point (so we deopt to before the original dispatch op and
run the dispatch afresh). This much we got correct. However, it is also
important the deopt usages of registers are preserved. In the case that
the translation of a dispatch program produces no guards, but we do
stack some up for the sake of running a particular specialization, we
could end up losing those deopt usages, and thus have optimizations take
place that led to broken deopts. Use the existing solution for this: add
a synthetic deopt index also, which is used in order to keep alive the
deopt usages of the original deopt index we cloned.