Elide proven arg type guards in DP translation
If we already know that the property asserted by the guard is true -
either through type analysis or an existing guard - we can avoid
inserting it at all. This is quite a lot cheaper than adding it and
having it removed later (although that doesn't seem to actually take
place at present, even if in principle it should), since we avoid doing
the whole SSA version split dance.