Fix frame walker confused when traversing a saved context
For the currently executing frame, the frame walker uses the program counter
directly to get the current position (for determining whether we are in an
inlined frame) rather than a deopt index as we might not be on an instruction
carrying a deopt annotation.

However, when applying traversals for an MVMContext, we must not use the
program counter, as the MVMContext would have captured a call stack on which
we were in a different position in the frame. In this case the deopt index is
still the correct way to check for inlines.

Fix by only doing the check for tc->cur_frame if we're just starting out to
traverse.

This fixes the case where we use callframe() to capture a context and use e.g.
EXISTS-KEY on it and the latter gets inlined into the same frame that called
callframe().