Add support for HLL guards; use in lang-hllize
Some callsites become polymorphic or even megamorphic in the case that
they only care about HLL, but we enforce it with a type guard. The Raku
multi dispatch planner ran into this, meaning that in any non-trivial
program that dispatches over many types, we'd end up with a full inline
cache site with all the costs of that. Introduce guards on HLL in order
that we can avoid this situation.