Relax existence requirement on DP lookup tables
We can install a lookup table at a callsite in order to handle
megamorphic dispatches. However, the mechanism that existed so far
always made lack of entry in the table a reason for the dispatch to be
rejected. This is not useful for writing dispatchers where lack of an
entry is acceptable. Dispatchers that want the current semantics can
obtain them simply by adding a guard.