Fix frame walker getting confused when OSR outdates positions
When an MVMContext is created, we snapshot the current callstack, i.e. record
the current positions within the frames in their extras. These positions are
relative to the currently executing bytecode. OSR however replaces this
bytecode with specialized one, invalidating the recorded positions. This can
lead to the frame walker making wrong decisions with regards to whether we were
in an inlinee or not.

Fix by disabling OSR on frames that are part of a snapshot.

Fixes MVM_SPESH_BLOCKING=1 MVM_SPESH_NODELAY=1 ./rakudo-m -Ilib t/spec/S11-modules/require.t