Don't copy facts in version split
It's possible in odd situations (created by NODELAY, where the stats are
liable to be dubious at best) to end up with facts being on a virtual
regsister that aren't a subset of those added by the guard. In that
case, we could end up with bogus combinations of the two. Just don't
copy them; it seems there's little to no win anyway.