Reinstate adding speculative return type guards
We were logging return types, however due to changes to how dispatch
works, were completely ignoring them in the specializer. Change that. We
need to do things a little differently, because dispatch is a varargs
op, so we can't just calculate the offset before it trivially. Instead,
use a post-log annotation. However, that can collide with a log
annotation on the next instruction, so we cheat a little (there's no
reason these bytecode locations have to actually match up with the ops,
they just have to correlate between the spesh logging and the graph
building).

We also do a small tweak to logging to make sure that we don't log
return values from dispatchers, since those are unrelated and end up
causing all kinds of confusion (leading to astronomical deopt rates!)

This source of type information is really important in some cases, with
some microbenchmarks showing a factor of 2 improvement (although in many
cases it's nothing to slight improvement).