Be more precise about OSR caller positions
There are two situations in which we set the caller info needed flag:
one when we throw an exception and want to produce a backtrace, and
another when we need to do context introspection. Only the latter is in
absolute need of accurate position information, and thus must poison
OSR. This, together with non-traversable contexts, lets us get OSR back
in various situations, including some common cases of I/O, fixing a
performance regression relative to `master`.