All bits in the union returned by `make_temp_reg` should be initialised
`make_temp_reg` returns `MVMSpeshOperand`, which is a union between various
types and structs, all of which are 64 bits or smaller.

Code in `optimize_bb_switch` evaluates `lit_i64 != -1` before calling
`MVM_spesh_manipulate_release_temp_reg`. The struct within the union returned
from `make_temp_reg` is only 48 bits, so take care to ensure that the other
16 bits are initialised, else valgrind (correctly) complains that
"Conditional jump or move depends on uninitialised value(s)".

(To be fair, technically the subsequent code shouldn't be reading the union
using a different member than it was written with. If compilers are going to
get exacting about this - ie infer that writing the `lit_i64` member can be
completely eliminated - then we will need to change the struct within the
union to be exactly 64 bits in total.)