Bail from template JIT prior to adding labels
The template JIT graph builder would first consider annotations, and
after that see if there was a template. In the event that there was not
one, this could lead to labels being produced, but never emitted, which
then led to a missing label and the the entire JIT of the frame failing.
Resolve this by first checking if there is a template, and only then
creating any labels.

This does mean that there is a discrepancy between when the template JIT
wishes to add labels and when the lego JIT wishes to, which is worth a
closer look, however even if that is figured out, it's still better not
to do work in the template JIT that will ultimately be a throwaway due
to a missing template.