Fix failing return type check in load-precompilation-file
The loadbytecode* ops do not just load the bytecode, but also run the
deserialization frames of the loaded comp units. These frames may trigger a
global deoptimization. In this case, the frame with the loadbytecode* op will
be marked for lazy deopt. This means that it will be deopted, when we return
to it (i.e. when loadbytecode* finishes).

When this happened however, rakudo died with a "Type check failed for return
value; expected CompUnit::Handle:D but got BOOTIO (BOOTIO)" message. This error
was thrown by the raku-rv-typecheck dispatcher. This dispatcher is run in the
original bytecode, after a raku-rv-decont dispatcher which follows the
loadbytecodefh op. The reason for the error is that when deopting after
returning from the loadbytecodefh op, we skip the raku-rv-decont call and jump
straight to the raku-rv-typecheck call. So the latter gets run with unrelated
contents in the argument registers.

The reason for skipping the raku-rv-decont dispatch is that when the rv
dispatchers get speshed away, their Deopt One After annotations get moved onto
the loadbytecodefh op. As the loadbytecodefh op does not carry its own deopt
annotation, the moved deopt annotation makes us treat it like the deopt
actually happened in raku-rv-decont and we move past it.

Fix by giving the loadbytecode* ops their own Deopt One annotations, so we
reliably continue directly after the op in case we need to deoptimize.