Allow larger-than-default callstack records
Some very large argument flattenings can result in us wanting to
allocate huge flattening records on the callstack. This is not at all an
efficient thing to do, but we should retain support for flattenings up
to the size we did before the new dispatch mechanism. At the same time,
we don't want the call stack regions to be made especially large; in
fact, the current sizing is probably a bit too big for programs with
lots of continuations taken. Thus, allow for oversized records to be
allocated, fixing the issue and giving us flexibility to adopt a smaller
default size in the future, if we see that it would be worthwhile.