Update resume states only after all guards
When we work through a dispatch program with resumptions, we may need to
update the dispatch state. When there are nested resumptions, we may end
up with multiple of those updates, and they may have guards between
them. This meant that an update could be performed even if the overall
dispatch would fail and fall back to recording a new dispatch program.
Make sure to only update them when the dispatch program is fully
accepted.