Experimentally allow args to next resumption
This will hopefully be helpful in getting `callwith` working fully. It
is heavily restricted (by existing checks): the capture must be derived
directly from the arguments given to the `callwith` and may have some
extra constants added. It is not allowed to have values from resume init
states. This, however, seems like it shall not be required - which is
good, since it would be a significant additional amount of complexity.