Implement dispatcher-track-how
Allows the creation of dispatch programs that read the meta-object of
some other tracked value. No guards are put in place, since every object
has a meta-object, and any dispatch program wanting to make the access
to the HOW a constant can already easily do so by guarding on the type
and then inserting the read HOW as a literal.