Add a way to request non-installation of a DP
In some cases, we may know that we are only recording a new dispatch
program because of a transient situation. For example, a megamorphic
by type method dispatcher may end up in record mode because the class in
question didn't have a cached flat method table yet. Once that situation
is resolved, an already installed dispatch program could start to match.
Thus provide a way for a dispatcher to indicate that it's not worth the
installation of the current dispatch program, so we can hopefully avoid
one cause of piling up a mass of programs at the callsite beyond the
point we realize it's megamorphic.