Fix segfault when GCing ReentrantMutex
We cannot clone a ReentrantMutex. In the event that we attempt to do
so, an object is allocated, but then not initialized. That lack of
initialization would lead to a NULL pointer dereference when we tried to
GC the lock. (We don't need to take care of this anywhere else, because
the clone never leaks out to userspace; an exception is thrown before it
can.)