Increase stack depth tracking accuracy
Stack depth is used in order to make decisions about what order to
specialize things in. Since we sample the interpreted program, we can
end up with missing data at some points, and further we update the
statistics in batches, and can have buffer ends being reached in all
kinds of situations. Tracking the stack depth by keeping it on the frame
rather than as a global property seems to lead to more accurate results
and thus more regularly reaching peak performance.