Ensure we always spesh log full type tuples
If when we log an entry we don't have enough space in the spesh log to
send all of the parameter types, then send the current log now, so we
can record the complete entry and parameter types in the new one. This
is because the spesh stats incorporation will, at log boundaries,
discard any incomplete type tuple it sees.