Fix segfault due to uninitialized entry_point in native calls
Cloning a NativeCall site could leave it with initialized lib_handle, but an
entry_point of NULL if the original wasn't built yet. The check for whether we
have to initialize tha callsite before we can proceed with the call however
only looks at lib_handle. Since we unconditionally created a lib_handle when
cloning (even if the original didn't have one), we wouldn't initialize
entry_point and end up with a segfault.

Fix by only creating a new lib_handle when the original we're cloning is
actually fully initialized.