Fix nested runloops leaving call stack in confused state
Certain actions (like return) unwind the call stack to the next record
representing a call frame. When returning from a nested runloop (e.g. a
NativeCall callback) however, we must only unwind the records added by
the code running in the nested runloop. Otherwise we can end up unwinding
records that are still needed by a currently running dispatch in the outer
runloop.

To fix this We add a new kind of callstack record representing the boundary of
a nested runloop and holding a reference to the outer frame (because return
needs to know the frame to return to). Also make sure that the outer frame's
return type and reference to the register into which to write the return
value are restored fully.