Ensure stable order of flattened named args
Hash randomization means that the iteration of hash keys varies from
hash to hash. In the context of a flattened hash of arguments, this
means that a given set of named arguments could appear in many orders
upon iteration. Previously, we built up the callsite arg names list in
whatever order we got from hash iteration. That means we're liable to
get a lot of misses in callsite interning, and thus misses in the inline
cache on callsites that might only really be monomorphic or somewhat
polymorphic. Resolve this by sorting the keys. Hash code is deemded good
enough, since this isn't a correctness issue, and a collision is hugely
unlikely at all, let alone in larger enough numbers to make significant
differences to program performance.