Avoid possible threading issues with mixins
It's possible that one thread is currently accessing an object while another
one mixes a role into it. As in the latter case we may replace the object's
body, the first thread could end up accessing a body that we just freed in the
thread performing the mixin. Better be safe and use the fixed size allocator
to allocate the replaced object body and free a previously existing one at a
safe point.