Don't duplicately set up fake temps
We allow resume init values from one resumption to be used in the next
one. This works out because while we only keep info around about the
resumption we are currently handling, the resume init arg will have been
loaded into a temporary at the right time, and then the dispatch program
can later reference that temporary. When we are producing a dispatch
program recording, we didn't really run the resulting dispatch program,
so fake up the temporaries. We erroneously did this for every mention of
the resume init arg, however, meaning we could overwrite the correct
value already stored with a bogus one later. Fix it by making sure we
only ever accept the first value for a given fake temporary.