Fix segfault when thread deopts while another tries to find a dynamic
Commit e94893f897495f3de16d4898279c87e356058acb fixed a possible segfault in
the frame walkercaused by a concurrent deopt in a different thread. The fix was
to use a local copy of the frame's spesh_cand pointer. However
MVM_spesh_deopt_find_inactive_frame_deopt_idx which is called by the same code
also accesses the frame's spesh_cand. So this function could still end up
dereferencing a NULL pointer if the frame's spesh_cand was cleared in the mean
time.

Fix by adding a spesh_cand argument to that function, so we can use our local
copy consistently.