Fix segfault in frame walker when other thread clears the frame's extra
Similar to another thread clearing a frame's spesh_cand leading to issues fixed
in commit 74c35679a3fba707ad695c53c28918aca30b95b5, it's also possible that
another thread clears a frame's extra, right after we checked it for non-NULL
and before we try to dereference it.

Fix by taking a local copy of the pointer before we check and dereference it.