Drop provably pointless sp_resumption ops
When we perform an inlining, and there is no way that it can deopt, that
further implies there's no way it could trigger a resumption. In that
case, we do not need the sp_resumption instructions. While they are very
cheap to run, they may cause usages of other instructions also, which
can lead to further instruction deletions. This can in turn bring more
code beneath inline limits, which is always helpful.

The moved check for whether an instruction may cause a deopt is done in
order to prevent sp_bindcomplete - which we always delete upon inline -
resulting in us mistakenly thinking that there is an inline. (This may
have caused other missed opportunities before now.)