Fix possible garbage value in addition
An error in base64 decoding some data could result in a garbage value
being used as the length of what was decoded. In the case of an error,
an exception was thrown so that garbage value wasn't actually used, but
assign the correct value anyway to be safe against any future changes.