Add missing facts to spesh link guard
When we want to perform specialization linking, we insert a guard on
the argument we have, and potentially then decont and check what is in
the container. However, we weren't placing facts as a result of adding
the guard on the container, thus leaving an sp_decont in the resulting
output that could have instead been lowered to something cheaper. Add
them.