Allocate ->work registers on the call stack
Thus saving us a pair of FSA calls to allocate and free the registers
for every frame invocation. While microbenchmarks where we manage to
inline everything will just get a bit more done while interpreting, a
more realistic program will benefit from this. For example, the Rakudo
build process completes in around 96.7% of the time it used to before
this change.

While this is mostly a straightforward transition, there are two things
worth noting:

1. Continuations that get freed without being invoked must now make
   sure to invalidate any ->work pointers. Since this happens in GC
   we need to make sure we follow any forwarders to update the
   correct copy of a heap frame.
2. OSR gets a new edge case: it may now fail due to there not being
   enough space in the current stack region. This is not expected to
   be a common problem, but there is logging for if this happens.