Allocate some frame environments on the callstack
Only do this for frames that live on the callstack rather than on the
heap. This is the case when we have a lexical environment but it is
never captured or the frame doesn't escape in other ways. When frames do
escape, we have to also move the environment out of the callstack and
onto the heap. We already do go to some effort to allocate on the heap
in the first place when a frame tends to get promoted there anyway, so
the amount of times this movement will be required is limited. This is
mostly of benefit to Raku rather than NQP code, since in NQP we can
quite often totally eliminate lexicals, whereas with Raku that isn't
always possible.