Prepare for special return records on callstack
This adds the infrastructure for handling special return data as a type
of callstack record, which will ultimately allow for the removal of the
explicit check if there is a special return at frame exit. It allows for
storing special return data on the callstack rather than having to
malloc/free it.