Don't use special return for finalizer running
For now, we'll handle this by checking if there's anything in the
finalizer queue each time we exit a frame. It's not ideal, in that it
costs something every frame exit, but it unblocks eliminating the
special return mechanism, which will in turn unlock removing many other
checks on frame exit, so it's a reasonable trade-off.