Eliminate redundant checks/logic during return
Thanks to having a callstack record type to indicate the boundary of a
nested runloop (used in NativeCall callbacks), we no longer need to
explicitly check for the threaded interpreter boundary anywhere. We can
return NULL from the callstack unwind function, upon encountering the
native call boundary, to indicate interpreter exit instead. This gets
rid of a bunch of logic in remove_one_frame, and a little logic in
exit_frame.