Fix NULL pointer results getting boxed after native calls
NativeCall returns a type object for NULL pointer results. When lowering
sp_runnativecall_o to sp_runnativecall_i to facilitate JIT compilation, we need
to preserve this behaviour. We therefore emit new ops to check the return value
for non-zero before boxing. Otherwise we just assign the return type object to
the result.