Log types of native routine's return values.
This would usually be handled by MVM_args_set_result_obj, but we cannot use
that as it would assume that the callee got its own frame on the callstack.

Note: in addition to these logs we seem to be logging Nil results. No idea
where those come from. Needs investigation.