Fix thinko in inline dispatch resumption search
We are doing a forward search through the resumptions (because they were
emitted in reverse - most nested dispatcher first - already at dispatch
program compilation time). Thus the index should be incremented to reach
the end, not decremented. Could lead to out-of-bounds reads or just an
attempt to resume the wrong dispatcher depending on how it came up.