Fix threading issues caused by misalignment in --valgrind builds
When MoarVM is configured with --valgrind, the fixed size allocator pre- and
postfixes allocated memory with a number of red zone bytes, so valgrind can
detect buffer over- and underruns. The number of red zone bytes was set to 4.
This means that allocated memory blocks would start misaligned and the same
would be true for pointers in arrays or structs in these blocks. On x86_64
writes of pointers to memory can only be assumed to be atomic when they are
properly aligned.

The non-alignment meant that pointers were written in two halfs. This could
lead to threads reading only half of a pointer with the other half being read
from uninitialized (usually zeroed) memory which would be followed by
segfaults.

Fix by taking alignment into consideration and using 8 bytes for the red zone.