Fix race conditions in JIT compiled dispatches
The dispatch system's inline cache entries contain both a function pointer to
call and function specific data for use in that call. Cache entries start out
pointing to the unlinked_dispatch entry singleton which contains a pointer to
the dispatch_initial function. This will run the dispatch program recording
phase resulting in a dispatch program. At the end this dispatch program is
installed into the cache. This is done by replacing the cache entry pointer
in the cache with one pointing at the new result.

When reading from the cache, care must be taken to only dereference the cache
entry pointer once. Multiple reads may have different results as threads could
be competing to run the initial dispatch and replace the pointer. This was not
followed by code generated by the JIT compiler leading to dispatch functions
getting called with data they did not understand.

Fix by actually simplifying the code by re-using the dereferencing result we
already got during argument setup.