Remove counter productive NULL checks in decont ops
Object registers must never contain NULL values, so there should be no reason
to check for NULL values when deconting. Historically bugs in other parts of
the code have caused NULL values to appear in registers and these checks have
worked around these bugs. But in truth they just hide them and make it more
difficult to debug, so better to remove them. This might cost some stability in
the short term, but should enable us to fix the underlying bugs for good.