Fix invalid pred chain caused by MVM_spesh_manipulate_split_BB_at
MVM_spesh_manipulate_split_BB_at splits a basic block such as the later part
becomes a new basic block. However it neglected to update the successors of the
original basic block to point to the newly formed one instead. This caused
spesh to blow up when inlining. Fix by replacing all references to the original
in the successors' pred arrays by references to the new one.