Properly extend int return values of native calls to 64 bits
Native functions leave it to their caller to extend values with a size of less
than a full register width (as the caller may alternatively simply access only
the relevant part of the register). So we need to do that as well before we
write the full register back into the VM register. The old native call JIT did
this right already, so lets just salvage the code.