Fix unwind missing a frame after special return
When we tried to add a special return handler onto the call stack and we had to
allocate a new callstack region for this call stack record, the unwind
following the execution of the special return handler missed a frame. This was
because when we called MVM_callstack_unwind, on top of the callstack there was
a region record instead of a frame. MVM_callstack_unwind however assumes that
there is a frame on top and so didn't actually unwind the frame.

Fix by cleaning up the callstack before running a special return or unwind
handler.