Fix "No frame found in callstack region" error
When a frame did not fit into the same callstack region as the continuation tag
continuationcontrol would end up in a "No frame found in callstack region"
panic.

Fix by continuing the search in the following regions. This is ok because
MVM_callstack_continuation_slice does not just splice off the one region it
returns, but the full chain following it.