Fix uninitialized register after PEA and inlining
The register in question should have been written to by a PHI node and carried
the appropriate deopt usage markers. The other side of the PHI (just one
version) didn't. So PEA optimized away the writer.

Fix by copying the dopt_users from the PHIs target to the PHI's sources when
rewriting returns. This way PEA knows that it has to keep the writer alive.