Teach native_ref_store to distinguish between signed and unsigned ints
To address all the uint issues we also need MVM_nativeref_write_lex_u,
MVM_nativeref_write_attribute_u and MVM_nativeref_write_multidim_u. However,
our current tests don't expose this, and we don't need them for the
rebootstrapping. Hence leave this code (and subtle bugs) as-is until we have
the large and intertangled changes done and stable, to reduce the size and risk
of those changes.