New assign_u op
For now correctly handles assigning unsigneds into native positional references.

To address all the uint issues we also need code_pair_store_u and
value_desc_cont_store_u. However, our current tests don't expose this, and we
don't need them for the rebootstrapping. Hence leave this code (and subtle
bugs) as-is until we have the large and intertangled changes done and stable,
to reduce the size and risk of those changes.