For an empty hash, report that at most 0 buckets are in use
`MVM_str_hash_kompromat` is a private function that abstracts the logic for
determining the number of buckets in use, which is often slightly smaller
than the number of buckets allocated. This value is needed to know where
hash iterators should start.

Previously for an hash with memory allocated for buckets but no items stored
it would return the (non-zero) number of buckets available to be used, which
whilst correct is not optimal, as iterators would end up traversing the
entire buckets. However, for a hash with only a control structure allocated
it would actually trigger an assertion failure, which is not correct - this
is a legitimate call to make; the calling code is not buggy.

Hence fix this bug by returning "0 buckets in use" for the case of "control
structure only", and (also) optimise the case of "buckets allocated but no
items" by returning "0 buckets in use" for this too.

This is easy to implement, because "control structure only" is signalled by
having both cur_items and max_items set to 0, whilst the regular "buckets
allocated but no items" state is cur_items 0 and max_items non-zero.