Skip sorting the keys when serializing empty hashes
To ensure that bytecode generation is repeatable even with hashes randomised
we need to sort the hash keys so that we always serialize in a consistent
order. In turn, this means we need to allocate temporary storage for an
array of keys, so that we can sort this. All this code is generic and
correctly handles the case of 0 keys (and 1 key), but all of it can be
trivially skipped for 0 keys. Hence do so.

We don't generate that many empty hashes, so this isn't a massive win. But
it's worth doing because the conditional test is computationally trivial and
the code is clear.