Fix possible access to fromspace when autoboxing return values
If the target frame (e.g. tc->cur_frame) is a heap frame that lives in the
nursery, the return value is a native value (e.g. return_i or a native call),
the caller expects an object and boxing happens to trigger a GC run, the
target frame could be moved before we dereference the target pointer to get
the return_value register. This would lead to a segfault with GC_DEBUG 3.

Fix by dereferencing target first instead of having the autobox(_int) macros
doing it after boxing.

Fixes GH #1621