Fix return register staying uninitialized on dispatch to void native functions
This got lost when turning nativecall_invoke into nativecall_dispatch. Even if
the native function itself doesn't return a result (i.e. void), we have to set
the call's result to something as in general the call will be initiated through
dispatch_o which is expected to return an object. NativeCall convention is to
return a type object in this case. So follow the convention