Fix uninitialized return register on translated calls to void native functions
We optimize calls to native functions without return values (i.e. void) to
sp_runnativecall_v. Nevertheless we have to initialize the original target
register with something (i.e. the return value type object), because the
register may be read after a deopt. This is all the more likely as a guard
on the return register would trigger a deopt if the register contains a NULL
value.

Fixes segfaults after deopts after void native calls.