Assure memory write order in native call race condition fix
Commit 207db4bb5b75e86466fc63ec23f7f0c58fc45ed4 changed the order of
initialization of native call sites so concurrently running code won't read
uninitialized data. However the compiler didn't know that this order was
important and was free to re-order when optimizing. Add a memory barrier to
force the compiler to do things in the order we want it to.

Thanks to zhuomingliang++ for the suggestion!