Fix possible access to fromspace after deopt materialization
We stored the newly allocated objects we create during materialization in an
array, for future reference so we could store the same objects into multiple
target registers if required. Allocating these objects however could trigger
a GC run which then missed the pointers in the materialized array. This would
then lead to such outdated pointers getting written to frame registers.

Fix by not storing object pointers in the materialized array. Instead store
the index of the register that already holds a pointer to the desired object.
Since frame registers are always updated by the GC, it's safe to read from
them.

Fixes GH #1628